Crop Water Balance Accounting

Calculates several parameters of the crop water balance. It also suggests when to irrigate.

Usage

CWB(
  Rain,
  ET0,
  AWC,
  Drz,
  Kc = NULL,
  Irrig = NULL,
  MAD = NULL,
  InitialD = 0,
  start.date
)

Arguments

Rain

A vector, 1-column matrix or data.frame with daily rainfall totals in millimetres.

ET0

A vector, 1-column matrix or data.frame with daily reference evapotranspiration in millimetres.

AWC

A vector, 1-column matrix or data.frame with the available water capacity of the soil, that is: the amount of water between field capacity and permanent wilting point in millimetre of water per metres of soil, must be greater than or equal to 0.

Drz

A vector, 1-column matrix or data.frame defining the root zone depth in metres.

Kc

A vector, 1-column matrix or data.frame defining the crop coefficient. If NULL its values are assumed to be 1.

Irrig

A vector, 1-column matrix or data.frame with net irrigation amount infiltrated into the soil for the current day in millimetres.

MAD

A vector, 1-column matrix or data.frame defining the management allowed depletion. Varies between 0 and 1.

InitialD

Single number defining in millimetres, the initial soil water deficit. It is used to start the water balance accounting. Default value is 0, which assumes the root zone is at the field capacity.

start.date

Date at which the accounting should start. Formats: “YYYY-MM-DD”, “YYYY/MM/DD”.

Value

A data.frame of water balance accounting, including the soil water deficit.

Examples

Tavg <- DataForCWB[,2]
Tmax <- DataForCWB[,3]
Tmin <- DataForCWB[,4]
Rn <- DataForCWB[,6]
WS <- DataForCWB[,7]
RH <- DataForCWB[,8]
G <- DataForCWB[,9]
ET0 <- ET0_PM(Tavg, Tmax, Tmin, Rn, RH, WS, G, Alt = 700)
Rain <- DataForCWB[,10]
Drz <- DataForCWB[,11]
AWC <- DataForCWB[,12]
MAD <- DataForCWB[,13]
Kc <- DataForCWB[,14]
Irrig <- DataForCWB[,15]
CWB(Rain = Rain, ET0 = ET0, AWC = AWC, Drz = Drz,
    Kc = Kc, Irrig = Irrig, MAD = MAD, start.date = "2023-11-23")
#>            DaysSeason Rain Irrig ET0 Kc WaterStressCoef_Ks ETc (P+Irrig)-ETc
#> 2023-11-23          1 45.5     0 2.4  1                1.0 2.4          43.0
#> 2023-11-24          2  0.3     0 4.2  1                1.0 4.2          -3.9
#> 2023-11-25          3  0.0     0 4.3  1                1.0 4.3          -4.3
#> 2023-11-26          4 11.4     0 3.7  1                1.0 3.7           7.8
#> 2023-11-27          5  0.3     0 4.8  1                1.0 4.8          -4.6
#> 2023-11-28          6  0.0     0 5.7  1                1.0 5.7          -5.7
#> 2023-11-29          7  0.0     0 5.8  1                0.9 5.8          -5.8
#> 2023-11-30          8  0.5     0 3.0  1                0.8 3.0          -2.5
#> 2023-12-01          9 35.8     0 3.6  1                1.0 3.6          32.2
#> 2023-12-02         10  0.0     0 3.9  1                1.0 3.9          -3.9
#> 2023-12-03         11 11.9     0 3.2  1                1.0 3.2           8.8
#> 2023-12-04         12 21.3     0 3.2  1                1.0 3.2          18.2
#> 2023-12-05         13  0.0     0 4.2  1                1.0 4.2          -4.2
#> 2023-12-06         14 10.9     0 3.9  1                1.0 3.9           7.0
#> 2023-12-07         15  3.3     0 3.7  1                1.0 3.7          -0.4
#> 2023-12-08         16  0.0     0 4.2  1                1.0 4.2          -4.2
#> 2023-12-09         17  0.0     0 4.6  1                1.0 4.6          -4.6
#> 2023-12-10         18  0.0     0 5.4  1                1.0 5.4          -5.4
#> 2023-12-11         19  0.0     0 5.0  1                1.0 5.0          -5.0
#> 2023-12-12         20 17.5     0 3.3  1                1.0 3.3          14.2
#> 2023-12-13         21  0.3     0 3.5  1                1.0 3.5          -3.2
#> 2023-12-14         22 32.8     0 2.5  1                1.0 2.5          30.3
#> 2023-12-15         23  5.3     0 2.6  1                1.0 2.6           2.8
#> 2023-12-16         24  0.8     0 2.9  1                1.0 2.9          -2.1
#> 2023-12-17         25  7.4     0 2.3  1                1.0 2.3           5.1
#> 2023-12-18         26  3.6     0 2.6  1                1.0 2.6           1.0
#> 2023-12-19         27  1.3     0 1.8  1                1.0 1.8          -0.5
#> 2023-12-20         28  1.8     0 4.2  1                1.0 4.2          -2.5
#> 2023-12-21         29  0.0     0 4.9  1                1.0 4.9          -4.9
#> 2023-12-22         30  0.0     0 4.2  1                1.0 4.2          -4.2
#> 2023-12-23         31 12.7     0 3.0  1                1.0 3.0           9.7
#> 2023-12-24         32 20.3     0 3.2  1                1.0 3.2          17.1
#> 2023-12-25         33  0.0     0 3.6  1                1.0 3.6          -3.6
#> 2023-12-26         34  0.0     0 2.9  1                1.0 2.9          -2.9
#> 2023-12-27         35 14.5     0 2.1  1                1.0 2.1          12.4
#> 2023-12-28         36  0.0     0 4.0  1                1.0 4.0          -4.0
#> 2023-12-29         37  0.0     0 5.1  1                1.0 5.1          -5.1
#> 2023-12-30         38  0.0     0 5.4  1                1.0 5.4          -5.4
#> 2023-12-31         39  0.0     0 4.9  1                1.0 4.9          -4.9
#> 2024-01-01         40  7.4     0 4.1  1                1.0 4.1           3.3
#> 2024-01-02         41 23.6     0 3.1  1                1.0 3.1          20.5
#> 2024-01-03         42  1.0     0 1.4  1                1.0 1.4          -0.4
#> 2024-01-04         43  0.0     0 1.3  1                1.0 1.3          -1.3
#> 2024-01-05         44  0.0     0 1.9  1                1.0 1.9          -1.9
#> 2024-01-06         45  0.0     0 3.0  1                1.0 3.0          -3.0
#> 2024-01-07         46 20.1     0 3.3  1                1.0 3.3          16.8
#> 2024-01-08         47  4.1     0 3.1  1                1.0 3.1           1.0
#> 2024-01-09         48  8.4     0 2.1  1                1.0 2.1           6.3
#> 2024-01-10         49 71.4     0 3.3  1                1.0 3.3          68.0
#> 2024-01-11         50  2.8     0 2.7  1                1.0 2.7           0.1
#> 2024-01-12         51  0.0     0 2.4  1                1.0 2.4          -2.4
#> 2024-01-13         52  0.0     0 2.1  1                1.0 2.1          -2.1
#> 2024-01-14         53  0.0     0 1.7  1                1.0 1.7          -1.7
#> 2024-01-15         54 38.9     0 2.7  1                1.0 2.7          36.1
#> 2024-01-16         55  1.8     0 3.9  1                1.0 3.9          -2.1
#> 2024-01-17         56  0.0     0 3.3  1                1.0 3.3          -3.3
#> 2024-01-18         57  0.0     0 2.1  1                1.0 2.1          -2.1
#> 2024-01-19         58  1.5     0 2.5  1                1.0 2.5          -1.0
#> 2024-01-20         59  0.0     0 2.7  1                1.0 2.7          -2.7
#> 2024-01-21         60 11.7     0 2.5  1                1.0 2.5           9.2
#> 2024-01-22         61  0.0     0 4.2  1                1.0 4.2          -4.2
#> 2024-01-23         62  0.0     0 3.5  1                1.0 3.5          -3.5
#> 2024-01-24         63  0.0     0 4.0  1                1.0 4.0          -4.0
#> 2024-01-25         64  0.0     0 4.4  1                1.0 4.4          -4.4
#> 2024-01-26         65  5.8     0 4.6  1                1.0 4.6           1.2
#> 2024-01-27         66  0.0     0 4.8  1                1.0 4.8          -4.8
#> 2024-01-28         67  0.0     0 4.4  1                1.0 4.4          -4.4
#> 2024-01-29         68  0.3     0 4.4  1                0.9 4.4          -4.1
#> 2024-01-30         69  0.0     0 3.5  1                0.9 3.5          -3.5
#> 2024-01-31         70  0.0     0 4.1  1                0.8 4.1          -4.1
#> 2024-02-01         71  3.8     0 3.0  1                0.8 3.0           0.8
#> 2024-02-02         72  0.3     0 2.8  1                0.8 2.8          -2.6
#> 2024-02-03         73  0.0     0 2.4  1                0.8 2.4          -2.4
#> 2024-02-04         74  5.6     0 1.5  1                0.8 1.5           4.1
#> 2024-02-05         75  6.3     0 2.0  1                0.9 2.0           4.3
#> 2024-02-06         76  0.0     0 4.2  1                0.8 4.2          -4.2
#> 2024-02-07         77  0.0     0 4.3  1                0.8 4.3          -4.3
#> 2024-02-08         78  0.0     0 4.0  1                0.7 4.0          -4.0
#> 2024-02-09         79  0.0     0 4.5  1                0.7 4.5          -4.5
#> 2024-02-10         80  0.0     0 4.1  1                0.7 4.1          -4.1
#> 2024-02-11         81  0.0     0 3.6  1                0.6 3.6          -3.6
#> 2024-02-12         82  1.8     0 4.1  1                0.6 4.1          -2.3
#> 2024-02-13         83  0.0     0 4.3  1                0.5 4.3          -4.3
#> 2024-02-14         84  0.0     0 3.2  1                0.5 3.2          -3.2
#> 2024-02-15         85  0.0     0 3.8  1                0.5 3.8          -3.8
#> 2024-02-16         86  0.0     0 3.1  1                0.4 3.1          -3.1
#> 2024-02-17         87 64.0     0 3.1  1                1.0 3.1          60.9
#> 2024-02-18         88 22.4     0 3.9  1                1.0 3.9          18.5
#> 2024-02-19         89  0.3     0 3.9  1                1.0 3.9          -3.6
#> 2024-02-20         90 15.5     0 4.5  1                1.0 4.5          11.0
#> 2024-02-21         91  0.0     0 4.4  1                1.0 4.4          -4.4
#> 2024-02-22         92 12.4     0 3.6  1                1.0 3.6           8.8
#> 2024-02-23         93 18.5     0 4.0  1                1.0 4.0          14.6
#> 2024-02-24         94  3.8     0 3.8  1                1.0 3.8           0.0
#> 2024-02-25         95  9.7     0 3.6  1                1.0 3.6           6.0
#> 2024-02-26         96  0.0     0 4.2  1                1.0 4.2          -4.2
#> 2024-02-27         97  0.0     0 3.9  1                1.0 3.9          -3.9
#> 2024-02-28         98  1.8     0 3.2  1                1.0 3.2          -1.5
#> 2024-02-29         99 36.3     0 1.6  1                1.0 1.6          34.7
#> 2024-03-01        100 29.0     0 1.9  1                1.0 1.9          27.1
#> 2024-03-02        101 13.5     0 1.4  1                1.0 1.4          12.0
#> 2024-03-03        102  7.4     0 1.6  1                1.0 1.6           5.8
#> 2024-03-04        103  9.7     0 1.8  1                1.0 1.8           7.9
#> 2024-03-05        104  8.4     0 2.3  1                1.0 2.3           6.1
#> 2024-03-06        105  0.3     0 2.7  1                1.0 2.7          -2.5
#> 2024-03-07        106  1.0     0 2.1  1                1.0 2.1          -1.1
#> 2024-03-08        107  0.8     0 2.5  1                1.0 2.5          -1.8
#> 2024-03-09        108  0.3     0 3.6  1                1.0 3.6          -3.3
#> 2024-03-10        109  0.0     0 3.5  1                1.0 3.5          -3.5
#> 2024-03-11        110  7.4     0 2.6  1                1.0 2.6           4.8
#> 2024-03-12        111 18.5     0 2.0  1                1.0 2.0          16.5
#> 2024-03-13        112  0.0     0 3.4  1                1.0 3.4          -3.4
#> 2024-03-14        113  0.0     0 3.9  1                1.0 3.9          -3.9
#> 2024-03-15        114  0.0     0 3.4  1                1.0 3.4          -3.4
#> 2024-03-16        115  0.0     0 3.7  1                1.0 3.7          -3.7
#> 2024-03-17        116  0.0     0 3.1  1                1.0 3.1          -3.1
#> 2024-03-18        117  0.8     0 2.7  1                1.0 2.7          -1.9
#> 2024-03-19        118  2.8     0 2.6  1                1.0 2.6           0.2
#> 2024-03-20        119  0.3     0 3.8  1                1.0 3.8          -3.6
#> 2024-03-21        120  0.0     0 3.8  1                1.0 3.8          -3.8
#> 2024-03-22        121  0.0     0 2.3  1                1.0 2.3          -2.3
#> 2024-03-23        122  0.0     0 2.6  1                1.0 2.6          -2.6
#> 2024-03-24        123  0.0     0 3.5  1                1.0 3.5          -3.5
#> 2024-03-25        124  0.0     0 4.0  1                1.0 4.0          -4.0
#> 2024-03-26        125  0.0     0 3.7  1                1.0 3.7          -3.7
#> 2024-03-27        126 18.0     0 3.6  1                1.0 3.6          14.4
#> 2024-03-28        127  0.0     0 3.1  1                1.0 3.1          -3.1
#> 2024-03-29        128 64.5     0 3.0  1                1.0 3.0          61.6
#> 2024-03-30        129  0.0     0 3.0  1                1.0 3.0          -3.0
#>            NonStandardCropEvap ET_Defict   TAW SoilWaterDeficit d_MAD
#> 2023-11-23                 2.4       0.0  45.7              0.0  13.7
#> 2023-11-24                 4.2       0.0  45.7              3.9  13.7
#> 2023-11-25                 4.3       0.0  45.7              8.2  13.7
#> 2023-11-26                 3.7       0.0  45.7              0.4  13.7
#> 2023-11-27                 4.8       0.0  45.7              5.0  13.7
#> 2023-11-28                 5.7       0.0  45.7             10.7  13.7
#> 2023-11-29                 5.3       0.5  45.7             16.5  13.7
#> 2023-11-30                 2.5       0.5  45.7             19.0  13.7
#> 2023-12-01                 3.6       0.0  76.2              0.0  22.9
#> 2023-12-02                 3.9       0.0  76.2              3.9  22.9
#> 2023-12-03                 3.2       0.0  76.2              0.0  22.9
#> 2023-12-04                 3.2       0.0  76.2              0.0  22.9
#> 2023-12-05                 4.2       0.0  76.2              4.2  22.9
#> 2023-12-06                 3.9       0.0  76.2              0.0  22.9
#> 2023-12-07                 3.7       0.0  76.2              0.4  22.9
#> 2023-12-08                 4.2       0.0  76.2              4.7  22.9
#> 2023-12-09                 4.6       0.0  76.2              9.2  22.9
#> 2023-12-10                 5.4       0.0  76.2             14.6  22.9
#> 2023-12-11                 5.0       0.0  76.2             19.6  22.9
#> 2023-12-12                 3.3       0.0  76.2              5.4  22.9
#> 2023-12-13                 3.5       0.0  76.2              8.6  22.9
#> 2023-12-14                 2.5       0.0  76.2              0.0  22.9
#> 2023-12-15                 2.6       0.0  76.2              0.0  22.9
#> 2023-12-16                 2.9       0.0  76.2              2.1  22.9
#> 2023-12-17                 2.3       0.0  76.2              0.0  22.9
#> 2023-12-18                 2.6       0.0  76.2              0.0  22.9
#> 2023-12-19                 1.8       0.0  76.2              0.5  22.9
#> 2023-12-20                 4.2       0.0  76.2              3.0  22.9
#> 2023-12-21                 4.9       0.0  76.2              7.9  22.9
#> 2023-12-22                 4.2       0.0  76.2             12.1  22.9
#> 2023-12-23                 3.0       0.0  76.2              2.5  22.9
#> 2023-12-24                 3.2       0.0  76.2              0.0  22.9
#> 2023-12-25                 3.6       0.0  76.2              3.6  22.9
#> 2023-12-26                 2.9       0.0  76.2              6.5  22.9
#> 2023-12-27                 2.1       0.0  76.2              0.0  22.9
#> 2023-12-28                 4.0       0.0  76.2              4.0  22.9
#> 2023-12-29                 5.1       0.0  76.2              9.1  22.9
#> 2023-12-30                 5.4       0.0  76.2             14.5  22.9
#> 2023-12-31                 4.9       0.0  76.2             19.4  22.9
#> 2024-01-01                 4.1       0.0 114.3             16.1  22.9
#> 2024-01-02                 3.1       0.0 114.3              0.0  22.9
#> 2024-01-03                 1.4       0.0 114.3              0.4  22.9
#> 2024-01-04                 1.3       0.0 114.3              1.7  22.9
#> 2024-01-05                 1.9       0.0 114.3              3.6  22.9
#> 2024-01-06                 3.0       0.0 114.3              6.6  22.9
#> 2024-01-07                 3.3       0.0 114.3              0.0  22.9
#> 2024-01-08                 3.1       0.0 114.3              0.0  22.9
#> 2024-01-09                 2.1       0.0 114.3              0.0  22.9
#> 2024-01-10                 3.3       0.0 114.3              0.0  22.9
#> 2024-01-11                 2.7       0.0 114.3              0.0  22.9
#> 2024-01-12                 2.4       0.0 114.3              2.4  22.9
#> 2024-01-13                 2.1       0.0 114.3              4.4  22.9
#> 2024-01-14                 1.7       0.0 114.3              6.2  22.9
#> 2024-01-15                 2.7       0.0 114.3              0.0  22.9
#> 2024-01-16                 3.9       0.0 114.3              2.1  22.9
#> 2024-01-17                 3.3       0.0 114.3              5.4  22.9
#> 2024-01-18                 2.1       0.0 114.3              7.5  22.9
#> 2024-01-19                 2.5       0.0 114.3              8.5  22.9
#> 2024-01-20                 2.7       0.0 114.3             11.2  22.9
#> 2024-01-21                 2.5       0.0 114.3              2.0  22.9
#> 2024-01-22                 4.2       0.0 114.3              6.2  22.9
#> 2024-01-23                 3.5       0.0 114.3              9.6  22.9
#> 2024-01-24                 4.0       0.0 114.3             13.6  22.9
#> 2024-01-25                 4.4       0.0 114.3             18.0  22.9
#> 2024-01-26                 4.6       0.0 114.3             16.8  22.9
#> 2024-01-27                 4.8       0.0 114.3             21.5  22.9
#> 2024-01-28                 4.3       0.2 114.3             26.0  22.9
#> 2024-01-29                 4.1       0.3 114.3             30.1  22.9
#> 2024-01-30                 3.1       0.4 114.3             33.6  22.9
#> 2024-01-31                 3.4       0.7 114.3             37.7  22.9
#> 2024-02-01                 2.5       0.5 114.3             36.9  22.9
#> 2024-02-02                 2.3       0.5 114.3             39.5  22.9
#> 2024-02-03                 1.9       0.5 114.3             41.9  22.9
#> 2024-02-04                 1.3       0.3 114.3             37.9  22.9
#> 2024-02-05                 1.8       0.2 114.3             33.5  22.9
#> 2024-02-06                 3.5       0.7 114.3             37.7  22.9
#> 2024-02-07                 3.4       0.9 114.3             42.0  22.9
#> 2024-02-08                 3.0       1.0 114.3             46.0  22.9
#> 2024-02-09                 3.1       1.3 114.3             50.5  22.9
#> 2024-02-10                 2.7       1.4 114.3             54.6  22.9
#> 2024-02-11                 2.2       1.4 114.3             58.2  22.9
#> 2024-02-12                 2.4       1.7 114.3             60.6  22.9
#> 2024-02-13                 2.3       2.0 114.3             64.8  22.9
#> 2024-02-14                 1.6       1.6 114.3             68.0  22.9
#> 2024-02-15                 1.8       2.0 114.3             71.8  22.9
#> 2024-02-16                 1.3       1.8 114.3             74.9  22.9
#> 2024-02-17                 3.1       0.0 114.3             14.0  22.9
#> 2024-02-18                 3.9       0.0 114.3              0.0  22.9
#> 2024-02-19                 3.9       0.0 114.3              3.6  22.9
#> 2024-02-20                 4.5       0.0 114.3              0.0  22.9
#> 2024-02-21                 4.4       0.0 114.3              4.4  22.9
#> 2024-02-22                 3.6       0.0 114.3              0.0  22.9
#> 2024-02-23                 4.0       0.0 114.3              0.0  22.9
#> 2024-02-24                 3.8       0.0 114.3              0.0  22.9
#> 2024-02-25                 3.6       0.0 114.3              0.0  22.9
#> 2024-02-26                 4.2       0.0 114.3              4.2  22.9
#> 2024-02-27                 3.9       0.0 114.3              8.1  22.9
#> 2024-02-28                 3.2       0.0 114.3              9.6  22.9
#> 2024-02-29                 1.6       0.0 114.3              0.0  45.7
#> 2024-03-01                 1.9       0.0 114.3              0.0  45.7
#> 2024-03-02                 1.4       0.0 114.3              0.0  45.7
#> 2024-03-03                 1.6       0.0 114.3              0.0  45.7
#> 2024-03-04                 1.8       0.0 114.3              0.0  45.7
#> 2024-03-05                 2.3       0.0 114.3              0.0  45.7
#> 2024-03-06                 2.7       0.0 114.3              2.5  45.7
#> 2024-03-07                 2.1       0.0 114.3              3.6  45.7
#> 2024-03-08                 2.5       0.0 114.3              5.3  45.7
#> 2024-03-09                 3.6       0.0 114.3              8.6  45.7
#> 2024-03-10                 3.5       0.0 114.3             12.2  45.7
#> 2024-03-11                 2.6       0.0 114.3              7.4  45.7
#> 2024-03-12                 2.0       0.0 114.3              0.0  45.7
#> 2024-03-13                 3.4       0.0 114.3              3.4  45.7
#> 2024-03-14                 3.9       0.0 114.3              7.3  45.7
#> 2024-03-15                 3.4       0.0 114.3             10.7  45.7
#> 2024-03-16                 3.7       0.0 114.3             14.4  45.7
#> 2024-03-17                 3.1       0.0 114.3             17.5  45.7
#> 2024-03-18                 2.7       0.0 114.3             19.5  45.7
#> 2024-03-19                 2.6       0.0 114.3             19.3  45.7
#> 2024-03-20                 3.8       0.0 114.3             22.8  45.7
#> 2024-03-21                 3.8       0.0 114.3             26.6  45.7
#> 2024-03-22                 2.3       0.0 114.3             28.9  45.7
#> 2024-03-23                 2.6       0.0 114.3             31.5  45.7
#> 2024-03-24                 3.5       0.0 114.3             34.9  45.7
#> 2024-03-25                 4.0       0.0 114.3             38.9  45.7
#> 2024-03-26                 3.7       0.0 114.3             42.6  45.7
#> 2024-03-27                 3.6       0.0 114.3             28.1  45.7
#> 2024-03-28                 3.1       0.0 114.3             31.2  45.7
#> 2024-03-29                 3.0       0.0 114.3              0.0  45.7
#> 2024-03-30                 3.0       0.0 114.3              3.0  45.7
#>                        D>=dmad
#> 2023-11-23                  No
#> 2023-11-24                  No
#> 2023-11-25                  No
#> 2023-11-26                  No
#> 2023-11-27                  No
#> 2023-11-28                  No
#> 2023-11-29 Yes. Irrigate 16 mm
#> 2023-11-30 Yes. Irrigate 19 mm
#> 2023-12-01                  No
#> 2023-12-02                  No
#> 2023-12-03                  No
#> 2023-12-04                  No
#> 2023-12-05                  No
#> 2023-12-06                  No
#> 2023-12-07                  No
#> 2023-12-08                  No
#> 2023-12-09                  No
#> 2023-12-10                  No
#> 2023-12-11                  No
#> 2023-12-12                  No
#> 2023-12-13                  No
#> 2023-12-14                  No
#> 2023-12-15                  No
#> 2023-12-16                  No
#> 2023-12-17                  No
#> 2023-12-18                  No
#> 2023-12-19                  No
#> 2023-12-20                  No
#> 2023-12-21                  No
#> 2023-12-22                  No
#> 2023-12-23                  No
#> 2023-12-24                  No
#> 2023-12-25                  No
#> 2023-12-26                  No
#> 2023-12-27                  No
#> 2023-12-28                  No
#> 2023-12-29                  No
#> 2023-12-30                  No
#> 2023-12-31                  No
#> 2024-01-01                  No
#> 2024-01-02                  No
#> 2024-01-03                  No
#> 2024-01-04                  No
#> 2024-01-05                  No
#> 2024-01-06                  No
#> 2024-01-07                  No
#> 2024-01-08                  No
#> 2024-01-09                  No
#> 2024-01-10                  No
#> 2024-01-11                  No
#> 2024-01-12                  No
#> 2024-01-13                  No
#> 2024-01-14                  No
#> 2024-01-15                  No
#> 2024-01-16                  No
#> 2024-01-17                  No
#> 2024-01-18                  No
#> 2024-01-19                  No
#> 2024-01-20                  No
#> 2024-01-21                  No
#> 2024-01-22                  No
#> 2024-01-23                  No
#> 2024-01-24                  No
#> 2024-01-25                  No
#> 2024-01-26                  No
#> 2024-01-27                  No
#> 2024-01-28 Yes. Irrigate 26 mm
#> 2024-01-29 Yes. Irrigate 30 mm
#> 2024-01-30 Yes. Irrigate 34 mm
#> 2024-01-31 Yes. Irrigate 38 mm
#> 2024-02-01 Yes. Irrigate 37 mm
#> 2024-02-02 Yes. Irrigate 39 mm
#> 2024-02-03 Yes. Irrigate 42 mm
#> 2024-02-04 Yes. Irrigate 38 mm
#> 2024-02-05 Yes. Irrigate 34 mm
#> 2024-02-06 Yes. Irrigate 38 mm
#> 2024-02-07 Yes. Irrigate 42 mm
#> 2024-02-08 Yes. Irrigate 46 mm
#> 2024-02-09 Yes. Irrigate 50 mm
#> 2024-02-10 Yes. Irrigate 55 mm
#> 2024-02-11 Yes. Irrigate 58 mm
#> 2024-02-12 Yes. Irrigate 61 mm
#> 2024-02-13 Yes. Irrigate 65 mm
#> 2024-02-14 Yes. Irrigate 68 mm
#> 2024-02-15 Yes. Irrigate 72 mm
#> 2024-02-16 Yes. Irrigate 75 mm
#> 2024-02-17                  No
#> 2024-02-18                  No
#> 2024-02-19                  No
#> 2024-02-20                  No
#> 2024-02-21                  No
#> 2024-02-22                  No
#> 2024-02-23                  No
#> 2024-02-24                  No
#> 2024-02-25                  No
#> 2024-02-26                  No
#> 2024-02-27                  No
#> 2024-02-28                  No
#> 2024-02-29                  No
#> 2024-03-01                  No
#> 2024-03-02                  No
#> 2024-03-03                  No
#> 2024-03-04                  No
#> 2024-03-05                  No
#> 2024-03-06                  No
#> 2024-03-07                  No
#> 2024-03-08                  No
#> 2024-03-09                  No
#> 2024-03-10                  No
#> 2024-03-11                  No
#> 2024-03-12                  No
#> 2024-03-13                  No
#> 2024-03-14                  No
#> 2024-03-15                  No
#> 2024-03-16                  No
#> 2024-03-17                  No
#> 2024-03-18                  No
#> 2024-03-19                  No
#> 2024-03-20                  No
#> 2024-03-21                  No
#> 2024-03-22                  No
#> 2024-03-23                  No
#> 2024-03-24                  No
#> 2024-03-25                  No
#> 2024-03-26                  No
#> 2024-03-27                  No
#> 2024-03-28                  No
#> 2024-03-29                  No
#> 2024-03-30                  No