Crop Water Balance Accounting With Fixed Time Periods for Irrigation

Calculates several parameters of the crop water balance. It also suggests how much irrigation to apply.

Usage

CWB_FixedSchedule(
  Rain,
  ET0,
  AWC,
  Drz,
  Kc = NULL,
  Irrig = NULL,
  MAD = NULL,
  InitialD = 0,
  Scheduling,
  start.date
)

Arguments

Rain

Vector, 1-column matrix or data frame with daily rainfall totals in millimetres.

ET0

Vector, 1-column matrix or data frame with daily reference evapotranspiration in millimetres.

AWC

Vector, 1-column matrix or data frame with the available water capacity of the soil, that is: the amount of water between field capacity and permanent wilting point in millimetres of water per metres of soil.

Drz

Vector, 1-column matrix or data frame defining the root zone depth in metres.

Kc

Vector, 1-column matrix or data frame defining the crop coefficient. If NULL its values are assumed to be 1.

Irrig

Vector, 1-column matrix or data frame with net irrigation amount infiltrated into the soil for the current day in millimetres.

MAD

Vector, 1-column matrix or data frame defining the management allowed depletion. Varies between 0 and 1.

InitialD

Single number defining in millimetre, the initial soil water deficit. It is used to start the water balance accounting. Default value is zero, which assumes the root zone is at the field capacity.

Scheduling

Single integer number defining the number of days between two consecutive irrigations.

start.date

Date at which the accounting should start. Formats: “YYYY-MM-DD”, “YYYY/MM/DD”.

Value

Water balance accounting, including the soil water deficit.

Examples

Tavg <- DataForCWB[, 2]
Tmax <- DataForCWB[, 3]
Tmin <- DataForCWB[, 4]
Rn <- DataForCWB[, 6]
WS <- DataForCWB[, 7]
RH <- DataForCWB[, 8]
G <- DataForCWB[, 9]
ET0 <- ET0_PM(Tavg, Tmax, Tmin, Rn, RH, WS, G, Alt = 700)
Rain <- DataForCWB[, 10]
Drz <- DataForCWB[, 11]
AWC <- DataForCWB[, 12]
MAD <- DataForCWB[, 13]
Kc <- DataForCWB[, 14]
Irrig <- DataForCWB[, 15]
Scheduling <- 5
CWB_FixedSchedule(
  Rain = Rain,
  ET0 = ET0,
  AWC = AWC,
  Drz = Drz,
  Kc = Kc,
  Irrig = Irrig,
  MAD = MAD,
  Scheduling = Scheduling,
  start.date = "2023-11-23"
)
#>            DaysSeason   Rain Irrig   ET0 Kc WaterStressCoef_Ks   ETc
#> 2023-11-23          1 45.470     0 2.440  1              1.000 2.440
#> 2023-11-24          2  0.254     0 4.172  1              1.000 4.172
#> 2023-11-25          3  0.000     0 4.290  1              1.000 4.290
#> 2023-11-26          4 11.430     0 3.665  1              1.000 3.665
#> 2023-11-27          5  0.254     0 4.849  1              1.000 4.849
#> 2023-11-28          6  0.000     0 5.670  1              1.000 5.670
#> 2023-11-29          7  0.000     0 5.757  1              0.914 5.757
#> 2023-11-30          8  0.508     0 3.009  1              0.836 3.009
#> 2023-12-01          9 35.810     0 3.588  1              1.000 3.588
#> 2023-12-02         10  0.000     0 3.929  1              1.000 3.929
#> 2023-12-03         11 11.940     0 3.179  1              1.000 3.179
#> 2023-12-04         12 21.340     0 3.180  1              1.000 3.180
#> 2023-12-05         13  0.000     0 4.239  1              1.000 4.239
#> 2023-12-06         14 10.920     0 3.925  1              1.000 3.925
#> 2023-12-07         15  3.302     0 3.737  1              1.000 3.737
#> 2023-12-08         16  0.000     0 4.228  1              1.000 4.228
#> 2023-12-09         17  0.000     0 4.572  1              1.000 4.572
#> 2023-12-10         18  0.000     0 5.381  1              1.000 5.381
#> 2023-12-11         19  0.000     0 4.968  1              1.000 4.968
#> 2023-12-12         20 17.530     0 3.340  1              1.000 3.340
#> 2023-12-13         21  0.254     0 3.460  1              1.000 3.460
#> 2023-12-14         22 32.770     0 2.477  1              1.000 2.477
#> 2023-12-15         23  5.334     0 2.564  1              1.000 2.564
#> 2023-12-16         24  0.762     0 2.869  1              1.000 2.869
#> 2023-12-17         25  7.366     0 2.271  1              1.000 2.271
#> 2023-12-18         26  3.556     0 2.588  1              1.000 2.588
#> 2023-12-19         27  1.270     0 1.785  1              1.000 1.785
#> 2023-12-20         28  1.778     0 4.231  1              1.000 4.231
#> 2023-12-21         29  0.000     0 4.928  1              1.000 4.928
#> 2023-12-22         30  0.000     0 4.237  1              1.000 4.237
#> 2023-12-23         31 12.700     0 3.032  1              1.000 3.032
#> 2023-12-24         32 20.320     0 3.233  1              1.000 3.233
#> 2023-12-25         33  0.000     0 3.625  1              1.000 3.625
#> 2023-12-26         34  0.000     0 2.877  1              1.000 2.877
#> 2023-12-27         35 14.480     0 2.067  1              1.000 2.067
#> 2023-12-28         36  0.000     0 3.990  1              1.000 3.990
#> 2023-12-29         37  0.000     0 5.102  1              1.000 5.102
#> 2023-12-30         38  0.000     0 5.396  1              1.000 5.396
#> 2023-12-31         39  0.000     0 4.932  1              1.000 4.932
#> 2024-01-01         40  7.366     0 4.090  1              1.000 4.090
#> 2024-01-02         41 23.620     0 3.103  1              1.000 3.103
#> 2024-01-03         42  1.016     0 1.441  1              1.000 1.441
#> 2024-01-04         43  0.000     0 1.256  1              1.000 1.256
#> 2024-01-05         44  0.000     0 1.917  1              1.000 1.917
#> 2024-01-06         45  0.000     0 2.979  1              1.000 2.979
#> 2024-01-07         46 20.070     0 3.274  1              1.000 3.274
#> 2024-01-08         47  4.064     0 3.054  1              1.000 3.054
#> 2024-01-09         48  8.380     0 2.121  1              1.000 2.121
#> 2024-01-10         49 71.370     0 3.338  1              1.000 3.338
#> 2024-01-11         50  2.794     0 2.674  1              1.000 2.674
#> 2024-01-12         51  0.000     0 2.383  1              1.000 2.383
#> 2024-01-13         52  0.000     0 2.062  1              1.000 2.062
#> 2024-01-14         53  0.000     0 1.744  1              1.000 1.744
#> 2024-01-15         54 38.860     0 2.720  1              1.000 2.720
#> 2024-01-16         55  1.778     0 3.894  1              1.000 3.894
#> 2024-01-17         56  0.000     0 3.331  1              1.000 3.331
#> 2024-01-18         57  0.000     0 2.080  1              1.000 2.080
#> 2024-01-19         58  1.524     0 2.523  1              1.000 2.523
#> 2024-01-20         59  0.000     0 2.660  1              1.000 2.660
#> 2024-01-21         60 11.680     0 2.479  1              1.000 2.479
#> 2024-01-22         61  0.000     0 4.177  1              1.000 4.177
#> 2024-01-23         62  0.000     0 3.471  1              1.000 3.471
#> 2024-01-24         63  0.000     0 3.954  1              1.000 3.954
#> 2024-01-25         64  0.000     0 4.447  1              1.000 4.447
#> 2024-01-26         65  5.842     0 4.606  1              1.000 4.606
#> 2024-01-27         66  0.000     0 4.753  1              1.000 4.753
#> 2024-01-28         67  0.000     0 4.428  1              0.966 4.428
#> 2024-01-29         68  0.254     0 4.403  1              0.921 4.403
#> 2024-01-30         69  0.000     0 3.483  1              0.882 3.483
#> 2024-01-31         70  0.000     0 4.093  1              0.838 4.093
#> 2024-02-01         71  3.810     0 3.003  1              0.847 3.003
#> 2024-02-02         72  0.254     0 2.825  1              0.818 2.825
#> 2024-02-03         73  0.000     0 2.444  1              0.792 2.444
#> 2024-02-04         74  5.588     0 1.537  1              0.836 1.537
#> 2024-02-05         75  6.350     0 2.033  1              0.883 2.033
#> 2024-02-06         76  0.000     0 4.172  1              0.838 4.172
#> 2024-02-07         77  0.000     0 4.314  1              0.790 4.314
#> 2024-02-08         78  0.000     0 4.011  1              0.747 4.011
#> 2024-02-09         79  0.000     0 4.459  1              0.698 4.459
#> 2024-02-10         80  0.000     0 4.131  1              0.653 4.131
#> 2024-02-11         81  0.000     0 3.605  1              0.613 3.605
#> 2024-02-12         82  1.778     0 4.117  1              0.588 4.117
#> 2024-02-13         83  0.000     0 4.266  1              0.541 4.266
#> 2024-02-14         84  0.000     0 3.191  1              0.506 3.191
#> 2024-02-15         85  0.000     0 3.769  1              0.465 3.769
#> 2024-02-16         86  0.000     0 3.100  1              0.431 3.100
#> 2024-02-17         87 64.010     0 3.077  1              1.000 3.077
#> 2024-02-18         88 22.350     0 3.861  1              1.000 3.861
#> 2024-02-19         89  0.254     0 3.891  1              1.000 3.891
#> 2024-02-20         90 15.490     0 4.482  1              1.000 4.482
#> 2024-02-21         91  0.000     0 4.360  1              1.000 4.360
#> 2024-02-22         92 12.450     0 3.612  1              1.000 3.612
#> 2024-02-23         93 18.540     0 3.975  1              1.000 3.975
#> 2024-02-24         94  3.810     0 3.772  1              1.000 3.772
#> 2024-02-25         95  9.650     0 3.629  1              1.000 3.629
#> 2024-02-26         96  0.000     0 4.197  1              1.000 4.197
#> 2024-02-27         97  0.000     0 3.937  1              1.000 3.937
#> 2024-02-28         98  1.778     0 3.247  1              1.000 3.247
#> 2024-02-29         99 36.320     0 1.582  1              1.000 1.582
#> 2024-03-01        100 28.960     0 1.893  1              1.000 1.893
#> 2024-03-02        101 13.460     0 1.424  1              1.000 1.424
#> 2024-03-03        102  7.366     0 1.594  1              1.000 1.594
#> 2024-03-04        103  9.650     0 1.798  1              1.000 1.798
#> 2024-03-05        104  8.380     0 2.269  1              1.000 2.269
#> 2024-03-06        105  0.254     0 2.740  1              1.000 2.740
#> 2024-03-07        106  1.016     0 2.108  1              1.000 2.108
#> 2024-03-08        107  0.762     0 2.514  1              1.000 2.514
#> 2024-03-09        108  0.254     0 3.558  1              1.000 3.558
#> 2024-03-10        109  0.000     0 3.546  1              1.000 3.546
#> 2024-03-11        110  7.366     0 2.570  1              1.000 2.570
#> 2024-03-12        111 18.540     0 2.004  1              1.000 2.004
#> 2024-03-13        112  0.000     0 3.359  1              1.000 3.359
#> 2024-03-14        113  0.000     0 3.943  1              1.000 3.943
#> 2024-03-15        114  0.000     0 3.391  1              1.000 3.391
#> 2024-03-16        115  0.000     0 3.728  1              1.000 3.728
#> 2024-03-17        116  0.000     0 3.096  1              1.000 3.096
#> 2024-03-18        117  0.762     0 2.705  1              1.000 2.705
#> 2024-03-19        118  2.794     0 2.601  1              1.000 2.601
#> 2024-03-20        119  0.254     0 3.811  1              1.000 3.811
#> 2024-03-21        120  0.000     0 3.794  1              1.000 3.794
#> 2024-03-22        121  0.000     0 2.281  1              1.000 2.281
#> 2024-03-23        122  0.000     0 2.581  1              1.000 2.581
#> 2024-03-24        123  0.000     0 3.461  1              1.000 3.461
#> 2024-03-25        124  0.000     0 3.959  1              1.000 3.959
#> 2024-03-26        125  0.000     0 3.652  1              1.000 3.652
#> 2024-03-27        126 18.030     0 3.618  1              1.000 3.618
#> 2024-03-28        127  0.000     0 3.080  1              1.000 3.080
#> 2024-03-29        128 64.520     0 2.963  1              1.000 2.963
#> 2024-03-30        129  0.000     0 3.036  1              1.000 3.036
#>            (P+Irrig)-ETc NonStandardCropEvap ET_Defict    TAW SoilWaterDeficit
#> 2023-11-23        43.030               2.440     0.000  45.72            0.000
#> 2023-11-24        -3.918               4.172     0.000  45.72            3.918
#> 2023-11-25        -4.290               4.290     0.000  45.72            8.208
#> 2023-11-26         7.765               3.665     0.000  45.72            0.444
#> 2023-11-27        -4.595               4.849     0.000  45.72            5.038
#> 2023-11-28        -5.670               5.670     0.000  45.72           10.708
#> 2023-11-29        -5.757               5.263     0.495  45.72           16.465
#> 2023-11-30        -2.501               2.516     0.494  45.72           18.967
#> 2023-12-01        32.222               3.588     0.000  76.20            0.000
#> 2023-12-02        -3.929               3.929     0.000  76.20            3.929
#> 2023-12-03         8.761               3.179     0.000  76.20            0.000
#> 2023-12-04        18.160               3.180     0.000  76.20            0.000
#> 2023-12-05        -4.239               4.239     0.000  76.20            4.239
#> 2023-12-06         6.995               3.925     0.000  76.20            0.000
#> 2023-12-07        -0.435               3.737     0.000  76.20            0.435
#> 2023-12-08        -4.228               4.228     0.000  76.20            4.663
#> 2023-12-09        -4.572               4.572     0.000  76.20            9.235
#> 2023-12-10        -5.381               5.381     0.000  76.20           14.616
#> 2023-12-11        -4.968               4.968     0.000  76.20           19.583
#> 2023-12-12        14.190               3.340     0.000  76.20            5.394
#> 2023-12-13        -3.206               3.460     0.000  76.20            8.600
#> 2023-12-14        30.293               2.477     0.000  76.20            0.000
#> 2023-12-15         2.770               2.564     0.000  76.20            0.000
#> 2023-12-16        -2.107               2.869     0.000  76.20            2.107
#> 2023-12-17         5.095               2.271     0.000  76.20            0.000
#> 2023-12-18         0.968               2.588     0.000  76.20            0.000
#> 2023-12-19        -0.515               1.785     0.000  76.20            0.515
#> 2023-12-20        -2.453               4.231     0.000  76.20            2.968
#> 2023-12-21        -4.928               4.928     0.000  76.20            7.897
#> 2023-12-22        -4.237               4.237     0.000  76.20           12.134
#> 2023-12-23         9.668               3.032     0.000  76.20            2.466
#> 2023-12-24        17.087               3.233     0.000  76.20            0.000
#> 2023-12-25        -3.625               3.625     0.000  76.20            3.625
#> 2023-12-26        -2.877               2.877     0.000  76.20            6.502
#> 2023-12-27        12.413               2.067     0.000  76.20            0.000
#> 2023-12-28        -3.990               3.990     0.000  76.20            3.990
#> 2023-12-29        -5.102               5.102     0.000  76.20            9.093
#> 2023-12-30        -5.396               5.396     0.000  76.20           14.489
#> 2023-12-31        -4.932               4.932     0.000  76.20           19.420
#> 2024-01-01         3.276               4.090     0.000 114.30           16.144
#> 2024-01-02        20.517               3.103     0.000 114.30            0.000
#> 2024-01-03        -0.425               1.441     0.000 114.30            0.425
#> 2024-01-04        -1.256               1.256     0.000 114.30            1.680
#> 2024-01-05        -1.917               1.917     0.000 114.30            3.597
#> 2024-01-06        -2.979               2.979     0.000 114.30            6.576
#> 2024-01-07        16.796               3.274     0.000 114.30            0.000
#> 2024-01-08         1.010               3.054     0.000 114.30            0.000
#> 2024-01-09         6.259               2.121     0.000 114.30            0.000
#> 2024-01-10        68.032               3.338     0.000 114.30            0.000
#> 2024-01-11         0.120               2.674     0.000 114.30            0.000
#> 2024-01-12        -2.383               2.383     0.000 114.30            2.383
#> 2024-01-13        -2.062               2.062     0.000 114.30            4.445
#> 2024-01-14        -1.744               1.744     0.000 114.30            6.189
#> 2024-01-15        36.140               2.720     0.000 114.30            0.000
#> 2024-01-16        -2.116               3.894     0.000 114.30            2.116
#> 2024-01-17        -3.331               3.331     0.000 114.30            5.447
#> 2024-01-18        -2.080               2.080     0.000 114.30            7.527
#> 2024-01-19        -0.999               2.523     0.000 114.30            8.525
#> 2024-01-20        -2.660               2.660     0.000 114.30           11.185
#> 2024-01-21         9.201               2.479     0.000 114.30            1.984
#> 2024-01-22        -4.177               4.177     0.000 114.30            6.161
#> 2024-01-23        -3.471               3.471     0.000 114.30            9.632
#> 2024-01-24        -3.954               3.954     0.000 114.30           13.586
#> 2024-01-25        -4.447               4.447     0.000 114.30           18.033
#> 2024-01-26         1.236               4.606     0.000 114.30           16.796
#> 2024-01-27        -4.753               4.753     0.000 114.30           21.549
#> 2024-01-28        -4.428               4.277     0.151 114.30           25.977
#> 2024-01-29        -4.149               4.053     0.350 114.30           30.126
#> 2024-01-30        -3.483               3.073     0.409 114.30           33.609
#> 2024-01-31        -4.093               3.429     0.664 114.30           37.702
#> 2024-02-01         0.807               2.542     0.461 114.30           36.895
#> 2024-02-02        -2.571               2.312     0.513 114.30           39.466
#> 2024-02-03        -2.444               1.934     0.509 114.30           41.909
#> 2024-02-04         4.051               1.285     0.252 114.30           37.859
#> 2024-02-05         4.317               1.796     0.237 114.30           33.542
#> 2024-02-06        -4.172               3.494     0.678 114.30           37.713
#> 2024-02-07        -4.314               3.410     0.904 114.30           42.027
#> 2024-02-08        -4.011               2.995     1.017 114.30           46.039
#> 2024-02-09        -4.459               3.111     1.348 114.30           50.497
#> 2024-02-10        -4.131               2.696     1.435 114.30           54.628
#> 2024-02-11        -3.605               2.211     1.395 114.30           58.233
#> 2024-02-12        -2.339               2.419     1.698 114.30           60.572
#> 2024-02-13        -4.266               2.307     1.958 114.30           64.838
#> 2024-02-14        -3.191               1.615     1.576 114.30           68.029
#> 2024-02-15        -3.769               1.752     2.017 114.30           71.798
#> 2024-02-16        -3.100               1.336     1.764 114.30           74.899
#> 2024-02-17        60.933               3.077     0.000 114.30           13.965
#> 2024-02-18        18.489               3.861     0.000 114.30            0.000
#> 2024-02-19        -3.637               3.891     0.000 114.30            3.637
#> 2024-02-20        11.008               4.482     0.000 114.30            0.000
#> 2024-02-21        -4.360               4.360     0.000 114.30            4.360
#> 2024-02-22         8.838               3.612     0.000 114.30            0.000
#> 2024-02-23        14.565               3.975     0.000 114.30            0.000
#> 2024-02-24         0.038               3.772     0.000 114.30            0.000
#> 2024-02-25         6.021               3.629     0.000 114.30            0.000
#> 2024-02-26        -4.197               4.197     0.000 114.30            4.197
#> 2024-02-27        -3.937               3.937     0.000 114.30            8.134
#> 2024-02-28        -1.469               3.247     0.000 114.30            9.603
#> 2024-02-29        34.738               1.582     0.000 114.30            0.000
#> 2024-03-01        27.067               1.893     0.000 114.30            0.000
#> 2024-03-02        12.036               1.424     0.000 114.30            0.000
#> 2024-03-03         5.772               1.594     0.000 114.30            0.000
#> 2024-03-04         7.852               1.798     0.000 114.30            0.000
#> 2024-03-05         6.111               2.269     0.000 114.30            0.000
#> 2024-03-06        -2.486               2.740     0.000 114.30            2.486
#> 2024-03-07        -1.092               2.108     0.000 114.30            3.578
#> 2024-03-08        -1.752               2.514     0.000 114.30            5.329
#> 2024-03-09        -3.304               3.558     0.000 114.30            8.633
#> 2024-03-10        -3.546               3.546     0.000 114.30           12.179
#> 2024-03-11         4.796               2.570     0.000 114.30            7.383
#> 2024-03-12        16.536               2.004     0.000 114.30            0.000
#> 2024-03-13        -3.359               3.359     0.000 114.30            3.359
#> 2024-03-14        -3.943               3.943     0.000 114.30            7.302
#> 2024-03-15        -3.391               3.391     0.000 114.30           10.694
#> 2024-03-16        -3.728               3.728     0.000 114.30           14.422
#> 2024-03-17        -3.096               3.096     0.000 114.30           17.518
#> 2024-03-18        -1.943               2.705     0.000 114.30           19.461
#> 2024-03-19         0.193               2.601     0.000 114.30           19.268
#> 2024-03-20        -3.557               3.811     0.000 114.30           22.825
#> 2024-03-21        -3.794               3.794     0.000 114.30           26.619
#> 2024-03-22        -2.281               2.281     0.000 114.30           28.900
#> 2024-03-23        -2.581               2.581     0.000 114.30           31.480
#> 2024-03-24        -3.461               3.461     0.000 114.30           34.942
#> 2024-03-25        -3.959               3.959     0.000 114.30           38.901
#> 2024-03-26        -3.652               3.652     0.000 114.30           42.553
#> 2024-03-27        14.412               3.618     0.000 114.30           28.141
#> 2024-03-28        -3.080               3.080     0.000 114.30           31.220
#> 2024-03-29        61.557               2.963     0.000 114.30            0.000
#> 2024-03-30        -3.036               3.036     0.000 114.30            3.036
#>             d_MAD             Scheduling
#> 2023-11-23 13.716                     No
#> 2023-11-24 13.716                     No
#> 2023-11-25 13.716                     No
#> 2023-11-26 13.716                     No
#> 2023-11-27 13.716  Time to Irrigate 5 mm
#> 2023-11-28 13.716                     No
#> 2023-11-29 13.716                     No
#> 2023-11-30 13.716                     No
#> 2023-12-01 22.860                     No
#> 2023-12-02 22.860  Time to Irrigate 4 mm
#> 2023-12-03 22.860                     No
#> 2023-12-04 22.860                     No
#> 2023-12-05 22.860                     No
#> 2023-12-06 22.860                     No
#> 2023-12-07 22.860  Time to Irrigate 0 mm
#> 2023-12-08 22.860                     No
#> 2023-12-09 22.860                     No
#> 2023-12-10 22.860                     No
#> 2023-12-11 22.860                     No
#> 2023-12-12 22.860  Time to Irrigate 5 mm
#> 2023-12-13 22.860                     No
#> 2023-12-14 22.860                     No
#> 2023-12-15 22.860                     No
#> 2023-12-16 22.860                     No
#> 2023-12-17 22.860  Time to Irrigate 0 mm
#> 2023-12-18 22.860                     No
#> 2023-12-19 22.860                     No
#> 2023-12-20 22.860                     No
#> 2023-12-21 22.860                     No
#> 2023-12-22 22.860 Time to Irrigate 12 mm
#> 2023-12-23 22.860                     No
#> 2023-12-24 22.860                     No
#> 2023-12-25 22.860                     No
#> 2023-12-26 22.860                     No
#> 2023-12-27 22.860  Time to Irrigate 0 mm
#> 2023-12-28 22.860                     No
#> 2023-12-29 22.860                     No
#> 2023-12-30 22.860                     No
#> 2023-12-31 22.860                     No
#> 2024-01-01 22.860 Time to Irrigate 16 mm
#> 2024-01-02 22.860                     No
#> 2024-01-03 22.860                     No
#> 2024-01-04 22.860                     No
#> 2024-01-05 22.860                     No
#> 2024-01-06 22.860  Time to Irrigate 7 mm
#> 2024-01-07 22.860                     No
#> 2024-01-08 22.860                     No
#> 2024-01-09 22.860                     No
#> 2024-01-10 22.860                     No
#> 2024-01-11 22.860  Time to Irrigate 0 mm
#> 2024-01-12 22.860                     No
#> 2024-01-13 22.860                     No
#> 2024-01-14 22.860                     No
#> 2024-01-15 22.860                     No
#> 2024-01-16 22.860  Time to Irrigate 2 mm
#> 2024-01-17 22.860                     No
#> 2024-01-18 22.860                     No
#> 2024-01-19 22.860                     No
#> 2024-01-20 22.860                     No
#> 2024-01-21 22.860  Time to Irrigate 2 mm
#> 2024-01-22 22.860                     No
#> 2024-01-23 22.860                     No
#> 2024-01-24 22.860                     No
#> 2024-01-25 22.860                     No
#> 2024-01-26 22.860 Time to Irrigate 17 mm
#> 2024-01-27 22.860                     No
#> 2024-01-28 22.860                     No
#> 2024-01-29 22.860                     No
#> 2024-01-30 22.860                     No
#> 2024-01-31 22.860 Time to Irrigate 38 mm
#> 2024-02-01 22.860                     No
#> 2024-02-02 22.860                     No
#> 2024-02-03 22.860                     No
#> 2024-02-04 22.860                     No
#> 2024-02-05 22.860 Time to Irrigate 34 mm
#> 2024-02-06 22.860                     No
#> 2024-02-07 22.860                     No
#> 2024-02-08 22.860                     No
#> 2024-02-09 22.860                     No
#> 2024-02-10 22.860 Time to Irrigate 55 mm
#> 2024-02-11 22.860                     No
#> 2024-02-12 22.860                     No
#> 2024-02-13 22.860                     No
#> 2024-02-14 22.860                     No
#> 2024-02-15 22.860 Time to Irrigate 72 mm
#> 2024-02-16 22.860                     No
#> 2024-02-17 22.860                     No
#> 2024-02-18 22.860                     No
#> 2024-02-19 22.860                     No
#> 2024-02-20 22.860  Time to Irrigate 0 mm
#> 2024-02-21 22.860                     No
#> 2024-02-22 22.860                     No
#> 2024-02-23 22.860                     No
#> 2024-02-24 22.860                     No
#> 2024-02-25 22.860  Time to Irrigate 0 mm
#> 2024-02-26 22.860                     No
#> 2024-02-27 22.860                     No
#> 2024-02-28 22.860                     No
#> 2024-02-29 45.720                     No
#> 2024-03-01 45.720  Time to Irrigate 0 mm
#> 2024-03-02 45.720                     No
#> 2024-03-03 45.720                     No
#> 2024-03-04 45.720                     No
#> 2024-03-05 45.720                     No
#> 2024-03-06 45.720  Time to Irrigate 2 mm
#> 2024-03-07 45.720                     No
#> 2024-03-08 45.720                     No
#> 2024-03-09 45.720                     No
#> 2024-03-10 45.720                     No
#> 2024-03-11 45.720  Time to Irrigate 7 mm
#> 2024-03-12 45.720                     No
#> 2024-03-13 45.720                     No
#> 2024-03-14 45.720                     No
#> 2024-03-15 45.720                     No
#> 2024-03-16 45.720 Time to Irrigate 14 mm
#> 2024-03-17 45.720                     No
#> 2024-03-18 45.720                     No
#> 2024-03-19 45.720                     No
#> 2024-03-20 45.720                     No
#> 2024-03-21 45.720 Time to Irrigate 27 mm
#> 2024-03-22 45.720                     No
#> 2024-03-23 45.720                     No
#> 2024-03-24 45.720                     No
#> 2024-03-25 45.720                     No
#> 2024-03-26 45.720 Time to Irrigate 43 mm
#> 2024-03-27 45.720                     No
#> 2024-03-28 45.720                     No
#> 2024-03-29 45.720                     No
#> 2024-03-30 45.720                     No