Data for Water Balance Accounting

Daily meteorological data from a weather station in Campinas, Brazil and other parameters required for calculating the crop water balance. The meteorological data belongs to the Agronomic Institute (IAC).

Usage

DataForCWB

Format

An object of class data.frame with 129 rows and 15 columns.

Details

@format ## DataForCWB A data frame with 15 columns and 129 rows:

date

date

tmed

Average air temperature in Celsius degrees

tmax

Maximum air temperature in Celsius degrees

tmin

Minimum air temperature in Celsius degrees

Ra

Extraterrestrial solar radiation in MJ m-2 day-1

Rn

Net radiation in MJ m-2 day-1

W

Wind speed in m s-1

RH

Relative Humidity in %

G

Soil Heat Flux in MJ m-2 day-1

Rain

Rain in millimetres

Drz

Depth of the root zone in metres

AWC

available water capacity (amount of water between field capacity and permanent wilting point) in millimetre of water per metre of soil

MAD

management allowed depletion (between 0 and 1)

Kc

Crop coefficient (between 0 and 1)

Irrig

Applied net irrigation in millimetres

@source http://www.ciiagro.org.br/