Descriptive Statistics for Weather Variables

Calculates descriptive statistics for rainfall, evapotranspiration, or other variables.

Usage

Descriptive(Sample)

Arguments

Sample

A vector, 1-column matrix or data frame with rainfall, evapotranspiration, or other variable.

Value

A dataframe with:

sample mean (Avg),

sample median (Med),

sample standard variation (SD)

sample standard Error (SE)

maximum value (MaxValue)

minimum value (MinValue)

frequency of zeros (FreqZero%)

Examples

Rain <- DataForCWB[, 10]
Descriptive(Sample = Rain)
#>   SampleSize  Avg  Med    SD   SE MaxValue MinValue FreqZero%
#> 1        129 6.53 0.25 13.06 1.15    71.37        0     48.06