Reference Evapotranspiration Using Hargreaves-Samani Method

Calculates daily reference evapotranspiration amounts using the Hargreaves-Samani method.

Usage

ET0_HS(Ra, Tavg, Tmax, Tmin)

Arguments

Ra

A vector, 1-column matrix or data.frame with extraterrestrial solar radiation in MJ m-2 day-1.

Tavg

A vector, 1-column matrix or data.frame column with daily average air temperature.

Tmax

A vector, 1-column matrix or data.frame with daily maximum air temperature in Celsius degrees.

Tmin

A vector, 1-column matrix or data.frame with daily minimum air temperature in Celsius degrees.

Value

A matrix of 1-column with the same length as `the input values with the daily potential evapotranspiration values in millimetres.

See also

Examples

# See `?DataForCWB` for more on this data set
Tavg <- DataForCWB[, 2]
Tmax <- DataForCWB[, 3]
Tmin <- DataForCWB[, 4]
Ra <- DataForCWB[, 5]
ET0_HS(Ra = Ra, Tavg = Tavg, Tmax = Tmax, Tmin = Tmin)
#>             ET0
#>   [1,] 4.703700
#>   [2,] 5.331592
#>   [3,] 5.664174
#>   [4,] 6.163377
#>   [5,] 5.291303
#>   [6,] 6.251883
#>   [7,] 6.729301
#>   [8,] 5.842178
#>   [9,] 5.402246
#>  [10,] 5.879287
#>  [11,] 5.114586
#>  [12,] 5.679834
#>  [13,] 5.938400
#>  [14,] 6.224412
#>  [15,] 5.222179
#>  [16,] 6.110026
#>  [17,] 5.647148
#>  [18,] 5.345193
#>  [19,] 6.224733
#>  [20,] 6.120729
#>  [21,] 5.918763
#>  [22,] 5.487768
#>  [23,] 2.502177
#>  [24,] 3.889903
#>  [25,] 4.645368
#>  [26,] 5.062661
#>  [27,] 3.650110
#>  [28,] 5.207263
#>  [29,] 6.193182
#>  [30,] 6.162937
#>  [31,] 5.476575
#>  [32,] 4.961276
#>  [33,] 5.380375
#>  [34,] 4.670873
#>  [35,] 4.616043
#>  [36,] 5.124276
#>  [37,] 5.400041
#>  [38,] 5.975927
#>  [39,] 5.275812
#>  [40,] 5.342574
#>  [41,] 5.106516
#>  [42,] 3.400841
#>  [43,] 2.533630
#>  [44,] 3.767255
#>  [45,] 5.213336
#>  [46,] 4.775999
#>  [47,] 3.273512
#>  [48,] 4.437918
#>  [49,] 4.718801
#>  [50,] 5.099593
#>  [51,] 3.846518
#>  [52,] 4.105348
#>  [53,] 3.073796
#>  [54,] 3.417059
#>  [55,] 4.713057
#>  [56,] 5.273933
#>  [57,] 3.225092
#>  [58,] 5.106130
#>  [59,] 4.381459
#>  [60,] 3.406934
#>  [61,] 5.546836
#>  [62,] 6.142076
#>  [63,] 5.923304
#>  [64,] 6.002144
#>  [65,] 5.530127
#>  [66,] 5.937510
#>  [67,] 6.091720
#>  [68,] 6.161030
#>  [69,] 6.250167
#>  [70,] 6.128848
#>  [71,] 6.128293
#>  [72,] 4.626581
#>  [73,] 4.647027
#>  [74,] 3.113991
#>  [75,] 4.535491
#>  [76,] 5.792464
#>  [77,] 6.144532
#>  [78,] 5.951522
#>  [79,] 5.837989
#>  [80,] 5.781079
#>  [81,] 5.492158
#>  [82,] 5.783120
#>  [83,] 5.984137
#>  [84,] 5.294486
#>  [85,] 5.284787
#>  [86,] 4.899781
#>  [87,] 5.443750
#>  [88,] 5.626722
#>  [89,] 5.150220
#>  [90,] 5.517872
#>  [91,] 5.668033
#>  [92,] 5.713170
#>  [93,] 5.429031
#>  [94,] 5.331041
#>  [95,] 5.041295
#>  [96,] 5.421722
#>  [97,] 5.454002
#>  [98,] 4.922040
#>  [99,] 3.265867
#> [100,] 3.283575
#> [101,] 2.272644
#> [102,] 2.249697
#> [103,] 2.500062
#> [104,] 2.439799
#> [105,] 3.735846
#> [106,] 3.096465
#> [107,] 3.317668
#> [108,] 3.670696
#> [109,] 4.877035
#> [110,] 4.366300
#> [111,] 3.107338
#> [112,] 3.399608
#> [113,] 4.227428
#> [114,] 3.881785
#> [115,] 3.125995
#> [116,] 4.336730
#> [117,] 4.141049
#> [118,] 2.908626
#> [119,] 3.318850
#> [120,] 4.214585
#> [121,] 3.315588
#> [122,] 3.675095
#> [123,] 4.080290
#> [124,] 4.712074
#> [125,] 4.842946
#> [126,] 4.601762
#> [127,] 4.103764
#> [128,] 4.804674
#> [129,] 3.162737