Reference Evapotranspiration Using the Penman and Monteith Method

Calculates daily reference evapotranspiration amounts using the Penman and Monteith method.

Usage

ET0_PM(Tavg, Tmax, Tmin, Rn, RH, WS, G = NULL, Alt)

Arguments

Tavg

A vector, 1-column matrix or data frame with daily average air temperature.

Tmax

A vector, 1-column matrix or data frame with daily maximum air temperature in Celsius degrees.

Tmin

A vector, 1-column matrix or data frame with daily minimum air temperature in Celsius degrees.

Rn

A vector, 1-column matrix or data frame with daily net radiation in \(MJ m-2 day-1\).

RH

A vector, 1-column matrix or data frame with daily relative Humidity in \%.

WS

A vector, 1-column matrix or data frame with daily wind speed in \(m s-1\).

G

Optional. A vector, 1-column matrix or data frame with daily soil heat flux in \(MJ m-2 day-1\). Default is NULL and if NULL it is assumed to be zero. May be provided by Soil_Heat_Flux

Alt

A single number defining the altitude at crop's location in metres.

Value

A matrix of daily reference evapotranspiration amounts in millimetres.

Examples

# See `?DataForCWB` for more on this data set
Tavg <- DataForCWB[, 2]
Tmax <- DataForCWB[, 3]
Tmin <- DataForCWB[, 4]
Rn <- DataForCWB[, 6]
WS <- DataForCWB[, 7]
RH <- DataForCWB[, 8]
G <- DataForCWB[, 9]
ET0_PM(Tavg = Tavg,
       Tmax = Tmax,
       Tmin = Tmin,
       Rn = Rn,
       RH = RH,
       WS = WS,
       G = G,
       Alt = 700)
#>          ET0_PM
#>   [1,] 2.440372
#>   [2,] 4.171917
#>   [3,] 4.290477
#>   [4,] 3.665459
#>   [5,] 4.848520
#>   [6,] 5.669878
#>   [7,] 5.757218
#>   [8,] 3.009252
#>   [9,] 3.588073
#>  [10,] 3.929134
#>  [11,] 3.178951
#>  [12,] 3.179903
#>  [13,] 4.238886
#>  [14,] 3.924919
#>  [15,] 3.736914
#>  [16,] 4.227748
#>  [17,] 4.572188
#>  [18,] 5.380983
#>  [19,] 4.967634
#>  [20,] 3.340085
#>  [21,] 3.460195
#>  [22,] 2.476797
#>  [23,] 2.563764
#>  [24,] 2.868665
#>  [25,] 2.270788
#>  [26,] 2.587888
#>  [27,] 1.785461
#>  [28,] 4.230850
#>  [29,] 4.928216
#>  [30,] 4.237106
#>  [31,] 3.031897
#>  [32,] 3.232840
#>  [33,] 3.625152
#>  [34,] 2.876980
#>  [35,] 2.066601
#>  [36,] 3.990253
#>  [37,] 5.102416
#>  [38,] 5.395921
#>  [39,] 4.931788
#>  [40,] 4.089696
#>  [41,] 3.102730
#>  [42,] 1.440549
#>  [43,] 1.255634
#>  [44,] 1.916837
#>  [45,] 2.978990
#>  [46,] 3.274274
#>  [47,] 3.053868
#>  [48,] 2.121164
#>  [49,] 3.337522
#>  [50,] 2.674261
#>  [51,] 2.382919
#>  [52,] 2.062186
#>  [53,] 1.743649
#>  [54,] 2.719902
#>  [55,] 3.893881
#>  [56,] 3.331255
#>  [57,] 2.079598
#>  [58,] 2.522511
#>  [59,] 2.659627
#>  [60,] 2.479182
#>  [61,] 4.177189
#>  [62,] 3.470901
#>  [63,] 3.953771
#>  [64,] 4.446788
#>  [65,] 4.605793
#>  [66,] 4.752566
#>  [67,] 4.428409
#>  [68,] 4.402738
#>  [69,] 3.482928
#>  [70,] 4.093004
#>  [71,] 3.002991
#>  [72,] 2.824771
#>  [73,] 2.443523
#>  [74,] 1.537260
#>  [75,] 2.033014
#>  [76,] 4.171718
#>  [77,] 4.313748
#>  [78,] 4.011396
#>  [79,] 4.458537
#>  [80,] 4.130693
#>  [81,] 3.605419
#>  [82,] 4.116782
#>  [83,] 4.265619
#>  [84,] 3.191173
#>  [85,] 3.769367
#>  [86,] 3.100477
#>  [87,] 3.076770
#>  [88,] 3.860978
#>  [89,] 3.891256
#>  [90,] 4.481529
#>  [91,] 4.359935
#>  [92,] 3.612482
#>  [93,] 3.974897
#>  [94,] 3.771984
#>  [95,] 3.629168
#>  [96,] 4.196585
#>  [97,] 3.937338
#>  [98,] 3.246649
#>  [99,] 1.581817
#> [100,] 1.892770
#> [101,] 1.423680
#> [102,] 1.594165
#> [103,] 1.797832
#> [104,] 2.269151
#> [105,] 2.739517
#> [106,] 2.108067
#> [107,] 2.513851
#> [108,] 3.557549
#> [109,] 3.546166
#> [110,] 2.570338
#> [111,] 2.004163
#> [112,] 3.359461
#> [113,] 3.942910
#> [114,] 3.391423
#> [115,] 3.728216
#> [116,] 3.095896
#> [117,] 2.705479
#> [118,] 2.600727
#> [119,] 3.810780
#> [120,] 3.793675
#> [121,] 2.281158
#> [122,] 2.580672
#> [123,] 3.461290
#> [124,] 3.959155
#> [125,] 3.652050
#> [126,] 3.617627
#> [127,] 3.079701
#> [128,] 2.963331
#> [129,] 3.036478