Reference Evapotranspiration Using the Preistley-Taylor Method

Calculates daily reference evapotranspiration amounts using the Priestley-Taylor method.

Usage

ET0_PT(Tavg, Rn, G = NULL, Coeff = 1.26)

Arguments

Tavg

A vector, 1-column matrix or data frame with daily average air temperature.

Rn

A vector, 1-column matrix or data frame with daily net radiation in \(MJ m-2 day-1\).

G

Optional. A vector, 1-column matrix or data frame with daily soil heat flux in \(MJ m-2 day-1\). May be provided by Soil_Heat_Flux

Coeff

Single number defining the Priestley and Taylor coefficient. Default is 1.26.

Value

A matrix object of the daily potential evapotranspiration values in millimetres.

Examples

# See `?DataForCWB` for more on this data set
Tavg <- DataForCWB[, 2]
Rn <- DataForCWB[, 6]
G <- DataForCWB[, 9]
ET0_PT(Tavg = Tavg, Rn = Rn, G = G)
#>          ET0_PT
#>   [1,] 3.432709
#>   [2,] 5.849554
#>   [3,] 6.432616
#>   [4,] 5.695334
#>   [5,] 7.023900
#>   [6,] 7.817355
#>   [7,] 7.969599
#>   [8,] 4.393576
#>   [9,] 4.653462
#>  [10,] 5.978814
#>  [11,] 5.304786
#>  [12,] 4.664253
#>  [13,] 6.553145
#>  [14,] 5.484314
#>  [15,] 5.201022
#>  [16,] 6.503054
#>  [17,] 5.715570
#>  [18,] 6.909315
#>  [19,] 7.804435
#>  [20,] 5.077747
#>  [21,] 5.845326
#>  [22,] 4.293102
#>  [23,] 3.053320
#>  [24,] 4.884104
#>  [25,] 3.575602
#>  [26,] 4.604058
#>  [27,] 2.698581
#>  [28,] 6.133369
#>  [29,] 7.351322
#>  [30,] 6.916084
#>  [31,] 4.948312
#>  [32,] 4.325298
#>  [33,] 5.846221
#>  [34,] 5.229740
#>  [35,] 2.328130
#>  [36,] 5.710129
#>  [37,] 7.402896
#>  [38,] 8.153433
#>  [39,] 7.228566
#>  [40,] 6.020548
#>  [41,] 5.591019
#>  [42,] 2.795372
#>  [43,] 2.261419
#>  [44,] 2.904630
#>  [45,] 5.422204
#>  [46,] 5.998695
#>  [47,] 6.019247
#>  [48,] 3.484288
#>  [49,] 6.013609
#>  [50,] 4.528488
#>  [51,] 4.286374
#>  [52,] 3.810094
#>  [53,] 3.028765
#>  [54,] 4.208284
#>  [55,] 6.741852
#>  [56,] 6.173091
#>  [57,] 3.640752
#>  [58,] 4.583939
#>  [59,] 4.847745
#>  [60,] 3.469756
#>  [61,] 6.278070
#>  [62,] 4.744446
#>  [63,] 6.003765
#>  [64,] 6.592318
#>  [65,] 7.002078
#>  [66,] 7.226482
#>  [67,] 6.885681
#>  [68,] 6.901872
#>  [69,] 5.215671
#>  [70,] 6.089763
#>  [71,] 4.930984
#>  [72,] 4.739992
#>  [73,] 3.877625
#>  [74,] 2.444163
#>  [75,] 2.963215
#>  [76,] 6.456451
#>  [77,] 6.784993
#>  [78,] 5.013275
#>  [79,] 6.513050
#>  [80,] 6.626130
#>  [81,] 5.490744
#>  [82,] 6.076198
#>  [83,] 6.764660
#>  [84,] 4.622510
#>  [85,] 5.203206
#>  [86,] 4.417313
#>  [87,] 4.838369
#>  [88,] 5.880765
#>  [89,] 5.952378
#>  [90,] 5.917262
#>  [91,] 6.395339
#>  [92,] 5.619474
#>  [93,] 6.358471
#>  [94,] 6.106976
#>  [95,] 5.079787
#>  [96,] 5.347358
#>  [97,] 6.061473
#>  [98,] 5.660362
#>  [99,] 2.487020
#> [100,] 2.801926
#> [101,] 2.136770
#> [102,] 2.528400
#> [103,] 2.625068
#> [104,] 2.252863
#> [105,] 4.098172
#> [106,] 2.746582
#> [107,] 2.650451
#> [108,] 4.489595
#> [109,] 5.007572
#> [110,] 4.113771
#> [111,] 2.417191
#> [112,] 5.397230
#> [113,] 5.296223
#> [114,] 3.467445
#> [115,] 5.578272
#> [116,] 4.592273
#> [117,] 3.672019
#> [118,] 2.864292
#> [119,] 5.063881
#> [120,] 5.547227
#> [121,] 2.988344
#> [122,] 3.795440
#> [123,] 5.320628
#> [124,] 5.897138
#> [125,] 5.316923
#> [126,] 5.149022
#> [127,] 4.480929
#> [128,] 4.105992
#> [129,] 3.861724